Refactor: move private store to separate file
